store/wizard: move package types to content subslice
These are domain types describing wizard content state, not
component-specific types. Colocating them with the content subslice
aligns with the existing pattern for other subslices.
