+++ src/store/slices/wizard/index.ts
@@ -4,7 +4,8 @@ export * from './types';
 
 // submodules
 export * from './cloud';
-export * from './compliance/';
+export * from './compliance';
+export * from './content';
 export * from './details';
 export * from './filesystem';
 export * from './output';
